recipe/windows-pythondir.patch: extend the Windows patch
The new pkg-config file prefixes a few variables with a `${pc_sysroot}`
variable, which seems to default to `/` even on Windows. This messed up some of
our paths which are given in "mixed" format (`C:/Users/....`), breaking the
libxcb build. We generally want paths in our Windows pkg-config files to be in
mixed mode, so the most straightforward solution is to just get rid of the
sysroot variable. I'm not sure what the point of it is, but it shouldn't be
relevant for our packages.