recipe/build.sh: add Windows/Python 3.12 workaround
As documented in the pull request, the m2-automake1.15 package includes a version
of py-compile which requires the imp module, which has been removed in Python 3.12.
It would be good to update that package to a newer one, which does not have this
problem, but that will take a little bit. But we can work around easily in the
meantime.